mid sentence
